+++ .tools/cpplint.bash
@@ -9,8 +9,7 @@ SAMPLES02_DIR=$SCRIPT_DIR/../samples/samples02
 SAMPLES03_DIR=$SCRIPT_DIR/../samples/samples03
 
 function cpplint_check () {
-    cpplint --filter=-build/c++11,-runtime/reference,-build/include_subdir --linelength=100 --extensions=hpp,cpp \
-    $1/include/* $1/include/dynamixel/* $1/src/* $1/src/dynamixel/* $1/test/*
+    cpplint --filter=-build/c++11,-runtime/reference,-build/include_subdir --linelength=100 --extensions=hpp,cpp $1/include/* $1/src/* $1/test/*
 }
 
 echo "cpplintを実行し、コードフォーマットをチェックします"
