+++ rt_manipulators_lib/include/dynamixel_xm.hpp
@@ -12,15 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#ifndef RT_MANIPULATORS_LIB_INCLUDE_DYNAMIXEL_XM_HPP_
+#define RT_MANIPULATORS_LIB_INCLUDE_DYNAMIXEL_XM_HPP_
 
-#include "dynamixel/dynamixel_xm540.hpp"
+#include "dynamixel_base.hpp"
 
+namespace dynamixel_xm {
 
-namespace dynamixel_xm540 {
+class DynamixelXM : public dynamixel_base::DynamixelBase  {
+ public:
+  explicit DynamixelXM(const uint8_t id);
+  bool write_torque_enable(const dynamixel_base::comm_t & comm, const bool enable);
+};
 
-DynamixelXM540::DynamixelXM540(const uint8_t id)
-  : dynamixel_xm::DynamixelXM(id) {
-  name_ = "XM540";
-}
+}  // namespace dynamixel_xm
 
-}  // namespace dynamixel_xm540
+#endif  // RT_MANIPULATORS_LIB_INCLUDE_DYNAMIXEL_XM_HPP_
