+++ rt_manipulators_lib/src/CMakeLists.txt
@@ -12,9 +12,9 @@ add_library(${library_name}
     kinematics.cpp
     kinematics_utils.cpp
     config_file_parser.cpp
-    dynamixel/dynamixel_xm.cpp
-    dynamixel/dynamixel_xm430.cpp
-    dynamixel/dynamixel_xm540.cpp
+    dynamixel_xm.cpp
+    dynamixel_xm430.cpp
+    dynamixel_xm540.cpp
 )
 set_target_properties(${library_name} PROPERTIES VERSION 1.1.0 SOVERSION 1)
